+++ src/rules/__tests__/no-done-callback.test.ts
@@ -477,6 +477,14 @@ ruleTester.run('no-done-callback', rule, {
           messageId: 'noDoneCallback',
           line: 1,
           column: 37,
+          suggestions: [
+            {
+              messageId: 'suggestWrappingInPromise',
+              data: { callback: 'done' },
+              output:
+                'test.each``("something", () => {return new Promise(done => { done(); })})',
+            },
+          ],
         },
       ],
     },
@@ -487,6 +495,14 @@ ruleTester.run('no-done-callback', rule, {
           messageId: 'noDoneCallback',
           line: 1,
           column: 35,
+          suggestions: [
+            {
+              messageId: 'suggestWrappingInPromise',
+              data: { callback: 'done' },
+              output:
+                'it.each``("something", () => {return new Promise(done => { done(); })})',
+            },
+          ],
         },
       ],
     },