+++ src/rules/max-expects.ts
@@ -10,9 +10,7 @@ export default createRule({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Enforces a maximum number assertion calls in a test body',
-      recommended: false,
     },
     messages: {
       exceededMaxAssertion: