+++ src/rules/no-export.ts
@@ -5,9 +5,7 @@ export default createRule({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Disallow using `exports` in files containing tests',
-      recommended: 'error',
     },
     messages: {
       unexpectedExport: `Do not export from a test file`,