+++ src/rules/no-identical-title.ts
@@ -21,9 +21,7 @@ export default createRule({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Disallow identical titles',
-      recommended: 'error',
     },
     messages: {
       multipleTestTitle: