+++ src/rules/no-mocks-import.ts
@@ -17,9 +17,7 @@ export default createRule({
   meta: {
     type: 'problem',
     docs: {
-      category: 'Best Practices',
       description: 'Disallow manually importing from `__mocks__`',
-      recommended: 'error',
     },
     messages: {
       noManualImport: `Mocks should not be manually imported from a ${mocksDirName} directory. Instead use \`jest.mock\` and import from the original module path`,