+++ src/rules/no-untyped-mock-factory.ts
@@ -21,10 +21,8 @@ export default createRule({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description:
         'Disallow using `jest.mock()` factories without an explicit type parameter',
-      recommended: false,
     },
     messages: {
       addTypeParameterToModuleMock:
@@ -38,7 +36,7 @@ export default createRule({
   create(context) {
     return {
       CallExpression(node: TSESTree.CallExpression): void {
-        const { callee, typeParameters } = node;
+        const { callee, typeArguments } = node;
 
         if (callee.type !== AST_NODE_TYPES.MemberExpression) {
           return;
@@ -55,7 +53,7 @@ export default createRule({
           const [nameNode, factoryNode] = node.arguments;
 
           const hasTypeParameter =
-            typeParameters !== undefined && typeParameters.params.length > 0;
+            typeArguments !== undefined && typeArguments.params.length > 0;
           const hasReturnType =
             isFunction(factoryNode) && factoryNode.returnType !== undefined;