+++ src/rules/prefer-hooks-on-top.ts
@@ -4,9 +4,7 @@ export default createRule({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Suggest having hooks before any test cases',
-      recommended: false,
     },
     messages: {
       noHookOnTop: 'Hooks should come before test cases',