+++ src/rules/require-top-level-describe.ts
@@ -15,10 +15,8 @@ export default createRule<
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description:
         'Require test cases and hooks to be inside a `describe` block',
-      recommended: false,
     },
     messages,
     type: 'suggestion',