+++ src/rules/valid-describe-callback.ts
@@ -23,9 +23,7 @@ export default createRule({
   meta: {
     type: 'problem',
     docs: {
-      category: 'Possible Errors',
       description: 'Enforce valid `describe()` callback',
-      recommended: 'error',
     },
     messages: {
       nameAndCallback: 'Describe requires name and callback arguments',