+++ src/rules/valid-expect.ts
@@ -115,9 +115,7 @@ export default createRule<[Options], MessageIds>({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Enforce valid `expect()` usage',
-      recommended: 'error',
     },
     messages: {
       tooManyArgs: 'Expect takes at most {{ amount }} argument{{ s }}',