+++ src/rules/valid-title.ts
@@ -74,7 +74,7 @@ const compileMatcherPatterns = (
 type CompiledMatcherAndMessage = [matcher: RegExp, message?: string];
 type MatcherAndMessage = [matcher: string, message?: string];
 
-const MatcherAndMessageSchema: JSONSchema.JSONSchema7 = {
+const MatcherAndMessageSchema: JSONSchema.JSONSchema4 = {
   type: 'array',
   items: { type: 'string' },
   minItems: 1,
@@ -114,9 +114,7 @@ export default createRule<[Options], MessageIds>({
   name: __filename,
   meta: {
     docs: {
-      category: 'Best Practices',
       description: 'Enforce valid titles',
-      recommended: 'error',
     },
     messages: {
       titleMustBeString: 'Title must be a string',
@@ -158,6 +156,7 @@ export default createRule<[Options], MessageIds>({
               MatcherAndMessageSchema,
               {
                 type: 'object',
+                // @ts-expect-error https://github.com/eslint/eslint/discussions/17573
                 propertyNames: { enum: ['describe', 'test', 'it'] },
                 additionalProperties: {
                   oneOf: [{ type: 'string' }, MatcherAndMessageSchema],